Add some of the missing KTextEditor interface methods
Only the methods that are used internally in KTextEditor
were added, so e.g. Application::closeDocuments() which
is used in some Kate plugins was not implemented.

With this commit, the vi-mode commands :q, :close, etc.
will now work properly, especially the command :only
which previously resulted to an infinite loop.

All implementations were based on the corresponding
method's behavior in Kate.

BUG: 441908