clang: Make sure completion priority is verbose enough
We are squeezing a 0-80 priority range into 0-10 one, all integers. This
makes it easy for us to lose information rather soon.

Since we are only setting the MatchQuality for those Clang claims to be
between 0-20 (i.e. under CCP_SuperCompletion), we can narrow our
conversion to 0-20 too.

This allows us to properly sort some items that clang was properly
sorting but we were unable to express correctly.